requirements.txt
aniso8601==9.0.1
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==7.1.2
colorama==0.4.4
Django==3.2.4
django-request==1.5.6
Flask==1.1.4
flask-restx==0.4.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
Werkzeug==1.0.1
